Various adjustments in visitor

+++ aql/src/main/java/org/ehrbase/aql/dto/AqlQuery.java
@@ -31,8 +31,8 @@ public class AqlQuery {
     private Containment from;
     private WhereCondition where;
     private List<OrderByExpression> orderBy;
-    private Integer limit;
-    private Integer offset;
+    private Long limit;
+    private Long offset;
 
     public SelectClause getSelect() {
         return this.select;
@@ -66,19 +66,19 @@ public class AqlQuery {
         this.orderBy = orderBy;
     }
 
-    public Integer getLimit() {
+    public Long getLimit() {
         return limit;
     }
 
-    public void setLimit(Integer limit) {
+    public void setLimit(Long limit) {
         this.limit = limit;
     }
 
-    public Integer getOffset() {
+    public Long getOffset() {
         return offset;
     }
 
-    public void setOffset(Integer offset) {
+    public void setOffset(Long offset) {
         this.offset = offset;
     }
